Clustering benchmark improved.
